kinda works but svg is broken
