+++ src/components/GameMap/CountryGroup/CountryGroup.tsx
@@ -1,4 +1,5 @@
-import { useState } from "react";
+import { useState, useRef, useEffect } from "react";
+import panzoom from "panzoom";
 import "./CountryGroup.css";
 
 const CountryGroup = ({
@@ -13,8 +14,29 @@ const CountryGroup = ({
     clickedCountriesHandler(countryName);
   };
 
+  const canvasRef = useRef(null);
+  const panzoomRef = useRef(null); // Create a ref to hold the panzoom instance
+
+  useEffect(() => {
+    const canvas = panzoom(canvasRef.current, {
+      autocenter: true,
+      maxZoom: 3,
+      minZoom: 0.15,
+      initialX: 7000,
+      initialY: 0,
+    });
+
+    panzoomRef.current = canvas;
+
+    return () => {
+      canvas.dispose();
+      document.body.style.overflow = "auto";
+    };
+  }, []);
+
   return (
     <g
+      ref={canvasRef}
       onClick={handleClick}
       key={`${countryName}`}
       className={
